rename FullScreenSidePanel to more descriptive of its purpose SidePanelModal
