+++ kolibri/core/assets/src/views/SidePanelModal/SidePanelResourcesList.vue
