+++ kolibri/plugins/learn/assets/src/views/LibraryPage/SidePanel.vue
@@ -19,7 +19,7 @@
     />
     <!-- The full screen side panel is used on smaller screens, and toggles as an overlay -->
     <!-- FullScreen is a container component, and then the EmbeddedSidePanel sits within -->
-    <FullScreenSidePanel
+    <SidePanelModal
       v-else-if="mobileSidePanelIsOpen"
       class="full-screen-side-panel"
       data-test="filters-side-panel"
@@ -59,7 +59,7 @@
         @cancel="currentCategory = null"
         @input="handleCategory"
       />
-    </FullScreenSidePanel>
+    </SidePanelModal>
 
     <!-- Category Search modal for large screens. On smaller screens, it is -->
     <!-- contained within the full screen search modal (different design) -->
@@ -82,7 +82,7 @@
 
   import { ref } from 'kolibri.lib.vueCompositionApi';
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
-  import FullScreenSidePanel from 'kolibri.coreVue.components.FullScreenSidePanel';
+  import SidePanelModal from 'kolibri.coreVue.components.SidePanelModal';
   import responsiveWindowMixin from 'kolibri.coreVue.mixins.responsiveWindowMixin';
   import CategorySearchModal from '../CategorySearchModal';
   import EmbeddedSidePanel from '../EmbeddedSidePanel';
@@ -92,7 +92,7 @@
     components: {
       CategorySearchModal,
       EmbeddedSidePanel,
-      FullScreenSidePanel,
+      SidePanelModal,
     },
     mixins: [commonCoreStrings, responsiveWindowMixin],
     /* eslint-disable-next-line no-unused-vars */
